Hw5 auth

+++ src/services/contacts.js
@@ -1,5 +1,5 @@
-import { ContactsCollection } from '../db/models/contact.js';
 import { SORT_ORDER } from '../constants/index.js';
+import { ContactsCollection } from '../db/models/contacts.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 
 export const getAllContacts = async ({
@@ -8,9 +8,11 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
+
   const contactsQuery = ContactsCollection.find({});
 
   if (filter.contactType) {
@@ -19,6 +21,9 @@ export const getAllContacts = async ({
   if (typeof filter.isFavourite === 'boolean') {
     contactsQuery.where('isFavourite').equals(filter.isFavourite);
   }
+
+  contactsQuery.where('userId').equals(userId);
+
   const [contactsCount, contacts] = await Promise.all([
     ContactsCollection.find({}).merge(contactsQuery).countDocuments(),
     contactsQuery
@@ -28,6 +33,7 @@ export const getAllContacts = async ({
       .sort({ [sortBy]: sortOrder })
       .exec(),
   ]);
+
   const paginationData = calculatePaginationData(contactsCount, perPage, page);
 
   return {
@@ -36,8 +42,8 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (contactId) => {
-  const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async ({ _id, userId }) => {
+  const contact = await ContactsCollection.findById({ _id, userId });
   return contact;
 };
 
@@ -46,9 +52,9 @@ export const createContact = async (payload) => {
   return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
-  const updatedContact = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId },
+export const updateContact = async ({ _id, userId }, payload, options = {}) => {
+  const rawResult = await ContactsCollection.findOneAndUpdate(
+    { _id, userId },
     payload,
     {
       new: true,
@@ -56,19 +62,14 @@ export const updateContact = async (contactId, payload, options = {}) => {
       ...options,
     },
   );
-
-  if (!updatedContact || !updatedContact.value) return null;
-
+  if (!rawResult || !rawResult.value) return null;
   return {
-    contact: updatedContact.value,
-    isNew: Boolean(updatedContact?.lastErrorObject?.upserted),
+    contact: rawResult.value,
+    isNew: Boolean(rawResult?.lastErrorObject?.upserted),
   };
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await ContactsCollection.findOneAndDelete({
-    _id: contactId,
-  });
-
+export const deleteContact = async ({ _id, userId }) => {
+  const contact = await ContactsCollection.findOneAndDelete({ _id, userId });
   return contact;
 };
